1.19.4 sources

+++ client/net/minecraft/world/item/enchantment/FrostWalkerEnchantment.java
@@ -39,10 +39,10 @@ public class FrostWalkerEnchantment extends Enchantment {
    public static void onEntityMoved(LivingEntity var0, Level var1, BlockPos var2, int var3) {
       if (var0.isOnGround()) {
          BlockState var4 = Blocks.FROSTED_ICE.defaultBlockState();
-         float var5 = (float)Math.min(16, 2 + var3);
+         int var5 = Math.min(16, 2 + var3);
          BlockPos.MutableBlockPos var6 = new BlockPos.MutableBlockPos();
 
-         for(BlockPos var8 : BlockPos.betweenClosed(var2.offset((double)(-var5), -1.0, (double)(-var5)), var2.offset((double)var5, -1.0, (double)var5))) {
+         for(BlockPos var8 : BlockPos.betweenClosed(var2.offset(-var5, -1, -var5), var2.offset(var5, -1, var5))) {
             if (var8.closerToCenterThan(var0.position(), (double)var5)) {
                var6.set(var8.getX(), var8.getY() + 1, var8.getZ());
                BlockState var9 = var1.getBlockState(var6);
